Working fill colors!
Problem was clash between SWFShape_setLine2() and SWFShape_setRightFillStyle().
I think the issue is that setLine2 expects that it is the only style applied to
a shape and so when a fill style is also applied, stuff starts breaking. The
fix was to use setLine().  This will have to be addressed in the future in
order to support all the graphics parameters such as lend and lmitre.